loop_resolution.py fixes
The finding of graph cycles by NetworkX has a non-deterministic
component of some sort, which can cause considerable run-to-run
variability in calculating values which are involved in these cycles.
Simply sorting the list of cycles as returned by NetworkX ensures that
these are presented in the same order each time, which fixes the
run-to-run variability.