Bugfix: market share limit floating point comparison
Due to floating point storage, a "100%" market share allocation can be
stored as (e.g.) 1.000000002. When there are no more market shares to
adjust, the case is dealt with by setting the allowable limits to the
range (0,1), which should find the sole 100% market share compliant, but
the (implicit) strict equality condition in the test can't handle the
floating point "close-enoughness", and we get stuck in a loop.

This commit adds a use of numpy's `isclose` method to additionally check
if the current market share "is" equal to either limiting value.